# speed up asdf plugin startup on mac