--- a/plugins/asdf/asdf.plugin.zsh
+++ b/plugins/asdf/asdf.plugin.zsh
@@ -4,7 +4,7 @@ ASDF_COMPLETIONS="$ASDF_DIR/completions"
 
 # If not found, check for Homebrew package
 if [[ ! -f "$ASDF_DIR/asdf.sh" || ! -f "$ASDF_COMPLETIONS/asdf.bash" ]] && (( $+commands[brew] )); then
-   ASDF_DIR="$(brew --prefix asdf)"
+   ASDF_DIR="$(brew --prefix)/opt/asdf"
    ASDF_COMPLETIONS="$ASDF_DIR/etc/bash_completion.d"
 fi