@ouven 2021-01-22
For an explanation: https://lists.zx2c4.com/pipermail/password-store/2018-October/003420.html