@orien 2021-12-06
I don't notice a difference on my mac:


#### Before
```
❯ time zsh -i -c exit
zsh -i -c exit  0.34s user 0.19s system 96% cpu 0.554 total
```
#### After
```
❯ time zsh -i -c exit
zsh -i -c exit  0.34s user 0.19s system 94% cpu 0.556 total
```

#### Info
```
❯ brew --version
Homebrew 3.3.6-77-gcb7f27d
Homebrew/homebrew-core (git revision e5e57c81800; last commit 2021-12-06)
Homebrew/homebrew-cask (git revision 446f1f5f33; last commit 2021-12-06)
```