@mcornella 2021-12-21
Does not improve speed significantly. Thanks for the contribution though!